+++ core/src/main/java/com/ray3k/particleparkpro/runnables/OpenRunnable.java
@@ -66,5 +66,6 @@ public class OpenRunnable implements Runnable {
         effectEmittersPanel.hidePopEmitterControls();
 
         UndoManager.clear();
+        updateWindowTitle();
     }
 }