+++ core/src/main/java/com/ray3k/particleparkpro/runnables/SaveRunnable.java
@@ -74,6 +74,9 @@ public class SaveRunnable implements Runnable {
             }
 
             effectEmittersPanel.hidePopEmitterControls();
+
+            unsavedChangesMade = false;
+            Core.updateWindowTitle();
         } else {
             saveAsRunnable.run();
         }