+++ core/src/main/java/com/ray3k/particleparkpro/widgets/poptables/PopTemplate.java
@@ -84,5 +84,8 @@ public class PopTemplate extends PopTable {
         emitterPropertiesPanel.populateScrollTable(null);
 
         UndoManager.clear();
+
+        openFileFileHandle = null;
+        updateWindowTitle();
     }
 }